refactor: change catalog page name for consistency
