+++ src/Router.tsx
@@ -10,7 +10,7 @@ import { useSelector } from 'react-redux';
 import type { RootState } from './store/store';
 import CircularProgress from '@mui/material/CircularProgress';
 import ProfilePage from './features/profile/ProfilePage';
-import Catalog from './features/catalog/Catalog';
+import CatalogPage from './features/catalog/CatalogPage';
 
 type AuthGuardProps = {
   children: ReactNode;
@@ -42,7 +42,7 @@ export default function Router() {
         <Routes>
           <Route element={<Layout />}>
             <Route path={Paths.HOME} element={<MainPage />} />
-            <Route path={Paths.CATALOG} element={<Catalog />} />
+            <Route path={Paths.CATALOG} element={<CatalogPage />} />
             <Route path={Paths.ABOUT} element={<div>About Page</div>} />
 
             <Route
